+++ experiments/standalone/Bucket/global_bucket_temporalmap.jl
@@ -96,7 +96,7 @@ function setup_prob(t0, tf, Δt)
     bucket_domain = ClimaLand.Domains.SphericalShell(;
         radius = FT(6.3781e6),
         depth = soil_depth,
-        nelements = (10, 10),
+        nelements = (10, 10), # this failed with (50,10)
         npolynomial = 1,
         dz_tuple = FT.((1.0, 0.05)),
     )